+++ go.mod
@@ -2,8 +2,4 @@ module github.com/cloudflare/circl
 
 go 1.12
 
-require (
-	github.com/armfazh/h2c-go-ref v0.0.0-20200603094401-c69caec07161
-	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
-	golang.org/x/sys v0.0.0-20200602225109-6fdc65e7d980
-)
+require golang.org/x/sys v0.0.0-20200602225109-6fdc65e7d980
